refactor: interface
