+++ packages/shared/lib/message/index.ts
@@ -1,2 +1,2 @@
 export * from './type';
-export * from './message';
+export * from './messaging';
