+++ pages/content-ui/src/app.tsx
@@ -37,7 +37,7 @@ export default function App() {
 }
 
 const searchWeather = async (search: string) => {
-  const results = await messaging.sendMessage('SearchWeather', { search });
+  const results = await messaging.send('SearchWeather', { search });
   const weatherInfo = results.at(0);
   if (!weatherInfo) {
     alert('No weather information found');
